~~再不发版没机会了~~

+++ src/MaaCore/Vision/Infrast/InfrastFacilityImageAnalyzer.h
@@ -32,7 +32,7 @@ namespace asst
                 return {};
             }
             else {
-                if (index < 0 || index >= iter->second.size()) {
+                if (index < 0 || static_cast<size_t>(index) >= iter->second.size()) {
                     return {};
                 }
                 else {
